requirements.txt
fastapi
uvicorn
motor
pydantic
google-generativeai